Adds backend api controller for gettin all requests for user
